client/requirements.txt
grpcio
googleapis-common-protos
redis
pandas==1.0.3